# Conflicts:
#	userservice/src/main/java/com/example/userservice/service/UserService.java
#	userservice/src/test/java/com/example/userservice/service/TestUserService.java

+++ userservice/src/main/java/com/example/userservice/service/UserService.java
@@ -48,6 +48,10 @@ public class UserService {
   }
 
   public void deleteCustomer(String id) {
+    Optional<User> userInDatabase = dataApi.getUserById(id);
+    if (userInDatabase.isEmpty()) {
+      throw new RuntimeException("User does not exist");
+    }
     dataApi.deleteUser(id);
   }
 }
